Fixed constructor eslint error and some payment.js errors.
