+++ utils/payments/payments.js
@@ -5,83 +5,7 @@ import { loadScript } from '../../scripts/aem.js';
 import buildForm from '../forms/forms.js';
 import { resetCart } from '../../pages/cart/cart.js';
 import { resetOrderForm } from '../orderForm/orderForm.js';
-
-function formatPhoneNumberToE164(phoneNumber, countryCode = '1') {
-  // Remove all non-numeric characters
-  const digits = phoneNumber.replace(/\D/g, '');
-
-  // Ensure the phone number has at least 10 digits (US standard)
-  if (digits.length < 10) {
-    throw new Error('Invalid phone number. Must contain at least 10 digits.');
-  }
-
-  // Add the country code prefix if not already included
-  if (digits.startsWith(countryCode)) {
-    return `+${digits}`; // Already in E.164 format
-  }
-  return `+${countryCode}${digits}`;
-}
-
-// https://developer.squareup.com/docs/build-basics/common-data-types/working-with-addresses
-class SquarePaymentAddress {
-  constructor(formData) {
-    this.address_line_1 = formData.address1;
-    this.address_line_2 = formData.address2;
-    this.first_name = formData.name;
-    this.locality = formData.city; // city
-    this.administrative_district_level_1 = formData.state; // state
-    this.postal_code = formData.zipcode;
-  }
-
-  build() {
-    return {
-      address_line_1: this.address_line_1,
-      address_line_2: this.address_line_2,
-      first_name: this.first_name,
-      locality: this.locality,
-      administrative_district_level_1: this.administrative_district_level_1,
-      postal_code: this.postal_code,
-    };
-  }
-}
-
-class SquarePayment {
-  constructor(orderData, formData, token) {
-    this.idempotency_key = orderData.idempotency_key;
-    this.source_id = token;
-    this.amount_money = orderData.order.total_money;
-    this.buyer_email_address = formData.email;
-    this.buyer_phone_number = formatPhoneNumberToE164(formData.phone);
-    this.location_id = orderData.order.location_id;
-    this.order_id = orderData.order.id;
-    this.initializeAddresses(formData);
-    // this.tip_money = {}
-  }
-
-  initializeAddresses(formData) {
-    if (formData.getItShipped) {
-      this.shipping_address = new SquarePaymentAddress(formData).build();
-      this.billing_address = new SquarePaymentAddress(formData).build();
-    } else {
-      this.shipping_address = null;
-      this.billing_address = null;
-    }
-  }
-
-  build() {
-    return {
-      idempotency_key: this.idempotency_key,
-      source_id: this.source_id,
-      amount_money: this.amount_money,
-      billing_address: this.billing_address,
-      buyer_email_address: this.buyer_email_address,
-      buyer_phone_number: this.buyer_phone_number,
-      location_id: this.location_id,
-      order_id: this.order_id,
-      shipping_address: this.shipping_address,
-    };
-  }
-}
+import { SquarePayment } from '../../constructors/constructors.js';
 
 async function createSquarePayment(token, orderData, element) {
   const env = getEnvironment();
